+++ resources/tahomalocalapid/tahomalocalapid.py
@@ -151,6 +151,11 @@ def loginTahoma():
 			'Content-Type': 'application/x-www-form-urlencoded'
 		}
 
+		logging.debug("loginTahoma - start - ********************")
+		logging.debug("  - url : %s", url)
+		logging.debug("  - data : %s", payload)        
+		logging.debug("loginTahoma - start - ********************")        
+
 		response = requests.request("POST", url, headers=headers, data=payload)
 
 		if response.status_code and (response.status_code == 200):
@@ -210,7 +215,7 @@ def availableToken():
 			json_data = response.json()
 			for item in json_data:
 				#logging.info(" token  : %s", item)
-				if (item['label'] == 'JeedomTahomaLocalApi_token' and item['scope'] == 'devmode'):
+				if ((item['label'] == 'JeedomTahomaLocalApi_token' or item['label'] == 'JeedomTahomaLocalApi_token_' + _uuid )and item['scope'] == 'devmode'):
 					deleteToken(item['uuid'])
 		else:
 			logging.error("Http code : %s", response.status_code)
@@ -266,7 +271,7 @@ def getDevicesList():
 			logging.error("Http code : %s", response.status_code)
 			logging.error("Response : %s", response.json())
 			logging.error("Response header : %s", response.headers)	
-			shutdown()	
+			#shutdown()	
 
 	except requests.exceptions.HTTPError as err:
 		logging.error("Error when retrieving tahoma devices list -> %s",err)
@@ -313,7 +318,7 @@ def validateToken():
 		}
 
 		payload=json.dumps({
-				"label": "JeedomTahomaLocalApi_token",				
+				"label": "JeedomTahomaLocalApi_token" + _uuid,				
 				"token": _tokenTahoma ,
 				"scope": "devmode"
 		})		
@@ -326,7 +331,7 @@ def validateToken():
 			logging.error("Response header : %s", response.headers)
 			shutdown()
 		else:
-			jeedom_com.send_change_immediate({'tahomaSession' : {'pinCode' : _pincode, 'token' : _tokenTahoma}})
+			jeedom_com.send_change_immediate({'tahomaSession' : {'pinCode' : _pincode, 'token' : _tokenTahoma,'uuid' : _uuid}})
 		
 
 	except requests.exceptions.HTTPError as err:
@@ -524,7 +529,7 @@ def execForceRefresh(deviceUrl):
 	try:
 		url = _ipBox +'/enduser-mobile-web/1/enduserAPI/exec/apply'
 
-		payload=json.dumps({"label":"advancedRefresh","actions": [{"commands": [{"name": "advancedRefresh", "parameters": []}],"deviceURL": deviceUrl}]})
+		payload=json.dumps({"label":"advancedRefresh","actions": [{"commands": [{"name": "advancedRefresh", "parameters": ["advanced"]}],"deviceURL": deviceUrl}]})
 		
 		headers = {
 			'Content-Type' : 'application/json',
@@ -626,6 +631,7 @@ parser.add_argument("--pswd", help="Password for local api Tahoma", type=str)
 parser.add_argument("--pincode", help="Tahoma pin code", type=str)
 parser.add_argument("--boxLocalIp", help="Tahoma IP", type=str)
 parser.add_argument("--tahoma_token", help="Tahoma token", type=str)
+parser.add_argument("--uuid", help="Plugin uuid", type=str)
 args = parser.parse_args()
 
 if args.device:
@@ -652,6 +658,8 @@ if args.boxLocalIp:
 	_ipBox='https://' + args.boxLocalIp + ':8443'
 if args.tahoma_token:
 	_tokenTahoma=args.tahoma_token
+if args.uuid:
+	_uuid=args.uuid
 
 _socket_port = int(_socket_port)
 
@@ -670,6 +678,7 @@ logging.info('User: %s', _user)
 logging.info('Pin ocde: %s', _pincode)
 logging.info('Box IP: %s', _ipBox)
 logging.info('Tahoma token: %s', _tokenTahoma)
+logging.info('Plugin uuid: %s', _uuid)
 logging.info('*-------------------------------------------------------------------------*')
 
 signal.signal(signal.SIGINT, handler)